# bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Coolector",
    "description": "API documentation with Swagger",
    "version": "v1"
  },
  "paths": {
...
    "/api/user/new": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "POST: api/user\r\nCreates a New User",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "schema": {
              "$ref": "#/definitions/UserCreateDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/api/user/login": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Authenticates a user by verifying the provided email and password.\r\nIf the credentials are valid, a JWT token is generated and returned.\r\nOtherwise, an Unauthorized response is sent.",
        "description": "This method is accessed via the following endpoint: \r\nPOST: api/user/login\r\n\r\nIt uses the CoolectorAPI.DTO.UserLoginDTO object to capture the email and password submitted by the user.\r\nThe M:CoolectorAPI.Controllers.UserController.GenerateJwtToken(System.Int32) method is called to generate a JWT token for the authenticated user.",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "The data transfer object containing the user's login credentials (email and password).",
            "schema": {
              "$ref": "#/definitions/UserLoginDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
...
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "relation": {
          "type": "string"
        },
        "companyCode": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "UserCreateDTO": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "relation": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "UserLoginDTO": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "additionalProperties": false
    }
  }
}